refactor: move code doc widget into page feature
