+++ src/pages/code-doc/features/CodeDocView/lib/tsAdapter.ts
