+++ src/pages/code-doc/features/CodeDocView/ui/Icon.tsx
