+++ src/pages/code-doc/features/CodeDocView/ui/MermaidDiagram.tsx
