+++ src/pages/code-doc/features/CodeDocView/ui/RichText.tsx
