+++ src/pages/code-doc/features/CodeDocView/ui/SymbolSection.tsx
